+++ src/editors/Flow/view/Components/FlowTopBar/hooks/useNodeStatusUpdate.js
@@ -2,7 +2,6 @@ import { useState, useRef, useMemo, useEffect, useCallback } from "react";
 import _isEqual from "lodash/isEqual";
 import { RobotManager } from "@mov-ai/mov-fe-lib-core";
 import i18n from "../../../../../../i18n/i18n";
-import { FLOW_VIEW_MODE } from "../../../Constants/constants";
 import { compareDocumentPaths } from "../../../Utils/utils";
 
 const DEBOUNCE_TIME = 600; // ms
@@ -15,9 +14,8 @@ const useNodeStatusUpdate = (props, robotSelected, viewMode) => {
     id,
     version,
     alert,
-    nodeStatusUpdated,
-    nodeCompleteStatusUpdated,
-    onStartStopFlow
+    onStartStopFlow,
+    mainInterface,
   } = props;
   // State hooks
   const [robotStatus, setRobotStatus] = useState({
@@ -29,7 +27,6 @@ const useNodeStatusUpdate = (props, robotSelected, viewMode) => {
   // Refs
   const selectedRobotRef = useRef(robotSelected);
   const debounceDeltaRef = useRef(Date.now());
-  const nodeStatusViewModeRef = useRef(FLOW_VIEW_MODE.default);
   const nodeStatusRef = useRef({});
   const allNodeStatusRef = useRef({});
   const lastMessage = useRef("");
@@ -132,9 +129,8 @@ const useNodeStatusUpdate = (props, robotSelected, viewMode) => {
     nodeStatusRef.current = emptyNodeStatus;
     allNodeStatusRef.current = emptyAllNodeStatus;
     // Send updates to canvas
-    nodeStatusUpdated(emptyNodeStatus);
-    nodeCompleteStatusUpdated(emptyAllNodeStatus);
-  }, [nodeCompleteStatusUpdated, nodeStatusUpdated]);
+    mainInterface?.resetAllNodeStatus();
+  }, [mainInterface]);
 
   /**
    * Format flow path
@@ -169,6 +165,7 @@ const useNodeStatusUpdate = (props, robotSelected, viewMode) => {
   const updateNodeStatus = useCallback(
     (key, targetValue, data, forceUpdate) => {
       if (
+        !mainInterface ||
         Date.now() - debounceDeltaRef.current <= DEBOUNCE_TIME &&
         !forceUpdate
       )
@@ -196,19 +193,12 @@ const useNodeStatusUpdate = (props, robotSelected, viewMode) => {
           false
         );
 
-        if (
-          (!_isEqual(allNodeStatusRef.current, allNodesStatus) &&
-            isFlowRunning(activeFlow)) ||
-          nodeStatusViewModeRef.current !== viewMode
-        ) {
-          nodeStatusRef.current = nodeStatus;
-          allNodeStatusRef.current = allNodesStatus;
-          nodeStatusViewModeRef.current = viewMode;
-          nodeStatusUpdated(allNodesStatus, { isOnline, activeFlow });
-        }
-
         activeFlow = isOnline ? robotStatusData.active_flow : "";
         onStartStopFlow(activeFlow);
+        mainInterface?.nodeStatusUpdated(running ? Object.entries(allNodesStatus).reduce(
+          (a, [key, val]) => ({ ...a, [mainInterface.id + "__" + key]: val }),
+          {}
+        ) : { [mainInterface.id]: 0 }, { isOnline, activeFlow });
       }
       // Robot doesn't have "Status" key in Redis
       else {
@@ -230,9 +220,9 @@ const useNodeStatusUpdate = (props, robotSelected, viewMode) => {
       getNodesToUpdate,
       isFlowRunning,
       isRobotOnline,
-      nodeStatusUpdated,
       onStartStopFlow,
       onlineAlert,
+      mainInterface,
       viewMode
     ]
   );